empezando con el administrador de viajeros. Listo el bloqueo, desbloqueo de viajeros

+++ src/dao/usuario/UsuarioDAO.java
@@ -2,7 +2,7 @@
 package dao.usuario;
 import dao.generic.GenericDAO;
 import model.usuario.Usuario;
-public interface UsuarioDAO extends GenericDAO<Usuario, Integer> {
+public interface UsuarioDAO extends GenericDAO<Usuario, Long> {
 	public boolean eliminarUsuario(Usuario u);
     public boolean existe(String email);
     public Usuario getUsuario(String email);
